rename: follow convention
- add underline on folder name due to docs
